+++ CHANGES.rst
@@ -72,11 +72,23 @@
   primarily intended to reduce the number of round-trips to the
   database. This is a step towards :issue:`281`.
 
+- On MySQL, set the connection timezone to be UTC. This is necessary
+  to get values consistent between ``UTC_TIMESTAMP``,
+  ``UNIX_TIMESTAMP``, ``FROM_UNIXTIME``, and Python's ``time.gmtime``,
+  as used for comparing TIDs.
+
 - Make ``RelStorage.pack()`` also accept a TID from the RelStorage
   database to pack to. The usual Unix timestamp form for choosing a
   pack time can be ambiguous in the event of multiple transactions
   within a very short period of time. This is mostly a concern for
   automated tests.
+
+- Make PyMySQL use the same precision as mysqlclient when sending
+  floating point parameters.
+
+- Automatically detect when MySQL stored procedures in the database
+  are out of date with the current source in this package and replace
+  them.
 
 3.0a5 (2019-07-11)
 ==================